sidl integration, more gnt implementation, and refactoring

+++ strata/stdc/stdlib/rand.c
@@ -6,8 +6,8 @@ static uint32_t next = 1;
 
 int rand(void)
 {
-    next = next * 1103515245 + 12345;
-    return (uint32_t)(next >> 16) & RAND_MAX;
+    next = (next * 1103515245) + 12345;
+    return (int)((uint32_t)(next >> 16) & RAND_MAX);
 }
 
 void srand(unsigned int seed)
